# routes/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "API iMock"
  license:
    name: MIT
host: 192.168.56.102:8080
basePath: "/api"
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
paths:
  /service/{campaign}:
    post:
      description: "Create a vitual service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/Service"
      responses:
        201:
          description: "Service create"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    put:
      description: "Update a vitual service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/Service"
      responses:
        200:
          description: "Service update"
          schema:
           "$ref": "#/definitions/Message"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    delete:
      description: "Delete a vitual service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/ServiceSearch"
      responses:
        200:
          description: "Service delete"
          schema:
           "$ref": "#/definitions/Message"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    get:
      description: "Retrieve informations of service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"      
        - name: path
          in: query
          required: true
          description: "Path of service"
          type: string
          default: '{"path":"/INBound/OUTBound/REST_POST"}'
      responses:
        200:
          description: "Service detail"
          schema:
           "$ref": "#/definitions/Service"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
  /services/{campaign}:
    get:
      description: "List of services"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"      
      responses:
        200:
          description: "List of services"
          schema:
           "$ref": "#/definitions/ServiceList"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /response/{campaign}:
    post:
      description: "Add a response to service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/Response"
      responses:
        201:
          description: "Response create"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    put:
      description: "Update a response"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/Response"
      responses:
        200:
          description: "Response update"
          schema:
           "$ref": "#/definitions/Message"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    delete:
      description: "Delete a response"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"
        - name: data
          in: body
          required: true
          description: "Payload"
          schema:
           "$ref": "#/definitions/ResponseSearch"
      responses:
        200:
          description: "Response delete"
          schema:
           "$ref": "#/definitions/Message"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
    get:
      description: "Retrieve informations of service"
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"      
        - name: path
          in: query
          required: true
          description: "Path of service"
          type: string
          default: '{"path":"/INBound/OUTBound/REST_POST"}'
        - name: search
          in: query
          required: true
          description: "Criteria"
          type: string
          default: '3'
      responses:
        200:
          description: "Response infos"
          schema:
           "$ref": "#/definitions/Response"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
  /responses/{campaign}:
    get:
      description: "List of response "
      parameters:
        - name: campaign
          in: path
          required: true
          description: "Name of campaign"
          type: string
          default: "mock"      
        - name: path
          in: query
          required: true
          description: "Path of service"
          type: string
          default: '{"path":"/INBound/OUTBound/SOAP"}'          
      responses:
        200:
          description: "List of response"
          schema:
           "$ref": "#/definitions/ResponseList"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"           
  /campaigns/{group}:
    get:
      description: "List of campaigns"
      parameters:
        - name: group
          in: path
          required: true
          description: "Name of group"
          type: string
          default: "imock"      
      responses:
        200:
          description: "List of campaigns"
          schema:
           "$ref": "#/definitions/CampaignList"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /groups:
    get:
      description: "List of groups"
      responses:
        200:
          description: "List of groups"
          schema:
           "$ref": "#/definitions/GroupList"
        404:
          description: "Empty list"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /tools/clean:
    post:
      description: "Flush database"
      responses:
        200:
          description: "Database flush"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /tools/purge:
    post:
      description: "Delete expire request response pair"
      responses:
        200:
          description: "Request response pair delete"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /tools/import:
    post:
      description: "Import database sets"
      parameters:
        - name: filter
          in: query
          required: true
          description: "File filter"
          type: string
          default: '{"data":"^00_.*\\.(yaml)$"}'         
      responses:
        200:
          description: "Database sets import"
          schema:
           "$ref": "#/definitions/Message"
        500:
          description: "Error system"
          schema:
           "$ref": "#/definitions/Message"
  /tools/version:
    get:
      description: "Version"
      responses:
        200:
          description: "Database flush"
          schema:
           "$ref": "#/definitions/Message"

definitions:
  Service:
    type: object
    properties:
      path:
        description: "Path of service"
        type: string
        default: "\/INBound/OUTBound/REST_POST"
      designation:
        description: "Name of virtual service"
        type: string
        default: "imock unit tests"
      group:
        description: "Name of group"
        type: string
        default: "imock"
      attribut:
        description: "Key use to search response"
        type: string
        default: "id"
      contentType:
        description: "Response content-type"
        type: string
        default: "application/json; charset=UTF-8"
      status:
        description: "Response status"
        type: integer
        default: 200
      response:
        description: "Response default"
        type: string
        default: '{"id":0,"first_name":"foo","last_name":"bar","Label":""}'
      real:
        description: "Activate proxy mode"
        type: integer
        default: 0
      producer:
        description: "Address of real service"
        type: string
        default: ""
      min:
        description: "Use to calculate time of response"
        type: integer
        format: int64
        default: 100
      max:
        description: "Use to calculate time of response"
        type: integer
        format: int64
        default: 200
  ServiceSearch:
    type: object
    properties:
      path:
        description: "Path of service"
        type: string
        default: "\/INBound/OUTBound/REST_POST"
  ServiceList:
    type: object
    properties:
      service:
        description: "Service name"
        type: array
        items:
          type: string
  Response:
    type: object
    properties:
      path:
        description: "Path of service"
        type: string
        default: "\/INBound/OUTBound/REST_POST/3"
      response:
        description: "Response"
        type: string
        default: '{"id":3,"first_name":"jules","last_name":"verne","Label":"HAL-<!\\d{9}!>"}'
  ResponseSearch:
    type: object
    properties:
      path:
        description: "Path of service"
        type: string
        default: "\/INBound/OUTBound/REST_POST/3"
  ResponseList:
    type: object
    properties:
      response:
        description: "Response name"
        type: array
        items:
          type: string
  GroupList:
    type: object
    properties:
      group:
        description: "Group name"
        type: array
        items:
          type: string
  CampaignList:
    type: object
    properties:
      campaign:
        description: "Campaign name"
        type: array
        items:
          type: string
  Message:
    type: object
    properties:
      status:
        description: "Status"
        type: integer
        format: int64
      code:
        description: "Code of message or error"
        type: string
      message:
        description: "Message or error"
        type: string